Resolved merge conflicts:
- EventReferenceDTO location (kept main's gwt/ui/shared location)
- Import statements updated to match main's refactored DTO locations
- Launch files: accepted main's changes
- Preserved management-console-specific createEventSeries method

+++ java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/simulator/streamlets/RectField.java
@@ -6,10 +6,10 @@ import com.google.gwt.json.client.JSONObject;
 import com.google.gwt.json.client.JSONParser;
 import com.google.gwt.maps.client.base.LatLng;
 import com.google.gwt.maps.client.base.LatLngBounds;
-import com.sap.sailing.domain.common.Position;
-import com.sap.sailing.domain.common.impl.DegreePosition;
 import com.sap.sailing.gwt.ui.client.shared.racemap.CoordinateSystem;
 import com.sap.sailing.gwt.ui.simulator.StreamletParameters;
+import com.sap.sse.common.Position;
+import com.sap.sse.common.impl.DegreePosition;
 
 public class RectField implements VectorField {
 
@@ -137,7 +137,7 @@ public class RectField implements VectorField {
     }
 
     @Override
-    public double getMotionScale(int zoomLevel) {
+    public double getMotionScale(double zoomLevel) {
         return this.motionFactor * Math.pow(1.7, Math.min(1.0, 6.0 - zoomLevel));
     }
